+++ agape/src/lib.rs
@@ -119,6 +119,16 @@ impl ApplicationHandler for App<'_> {
                     .unwrap()
                     .resize_surface(size.width, size.height)
                     .expect("Failed to resize the pixel buffer");
+
+                self.pixels
+                    .as_mut()
+                    .unwrap()
+                    .resize_buffer(size.width, size.height)
+                    .expect("Failed to resize the pixel buffer");
+
+                let pixmap = Pixmap::new(size.width, size.height).unwrap();
+                self.pixmap = Some(pixmap);
+                self.resources.get_mut::<WindowSize>().unwrap().0 = Size::from(size);
             }
             WindowEvent::MouseInput { .. } => {}
             _ => {}
@@ -183,7 +193,6 @@ impl App<'_> {
         let widget = self.resources.get::<Box<dyn Widget>>().unwrap();
         let mut views: Vec<Box<dyn View>> = widget.iter().map(|w| w.view()).collect();
         let layout = self.resources.get::<Box<dyn Layout>>().unwrap();
-        // let mut views = &mut self.context.views;
 
         let pixels = self.pixels.as_mut().unwrap();
         let pixmap = self.pixmap.as_mut().unwrap();
@@ -191,15 +200,13 @@ impl App<'_> {
 
         // Draw each view(widget) to the pixmap
         for view in &mut views {
-            // TODO don't unwrap
             let layout = layout.get(view.id()).unwrap();
             view.set_size(layout.size());
             view.set_position(layout.position());
             view.render(pixmap);
-            // TODO copy data after
-            pixels.frame_mut().copy_from_slice(pixmap.data());
         }
 
+        pixels.frame_mut().copy_from_slice(pixmap.data());
         pixels.render().unwrap();
     }
 
@@ -214,6 +221,7 @@ impl App<'_> {
             .add_system(update_cursor_position)
             .add_system(handle_mouse_button)
             .add_system(intersection_observer)
+            .add_system(handle_key_input)
             .add_system(handle_widget_event);
 
         let event_loop = EventLoop::new()?;
@@ -267,6 +275,13 @@ fn handle_mouse_button(resources: &mut Resources, event: &WindowEvent) {
     }
 }
 
+fn handle_key_input(resources: &mut Resources, event: &WindowEvent) {
+    if let WindowEvent::KeyboardInput { event, .. } = event {
+        let events = resources.get_mut::<Vec<WidgetEvent>>().unwrap();
+        events.push(WidgetEvent::KeyInput(event.clone()));
+    }
+}
+
 fn intersection_observer(resources: &mut Resources) {
     let cursor_pos = resources.get::<CursorPosition>().unwrap();
     let layout = resources.get::<Box<dyn Layout>>().unwrap();
@@ -310,7 +325,7 @@ fn handle_widget_event(resources: &mut Resources) {
     let widget: &mut Box<dyn Widget> = resources.get_mut().unwrap();
 
     for event in events {
-        widget.handle_event(event);
+        widget.handle_event(&event);
     }
 
     resources.get_mut::<Vec<WidgetEvent>>().unwrap().clear();