+++ CHANGELOG.md
@@ -6,3 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+- Added `Resources` struct to keep share global resources
+
+### Changed
+- Systems now have a `&mut Resources` instead of the previous `&mut Context`