+++ CHANGELOG.md
@@ -8,12 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 0.2.0
 
 ### Added
+
 - Added `Resources` struct to share global resources.
 - Added repeat syntax (`hstack![widget;10]`) to `hstack!` and `vstack!` macros.
+- Added `on_click` and `on_hover` gestures.
 
 ### Changed
+
 - Systems now have a `&mut Resources` instead of the previous `&mut Context`.
-- Most of the functionality, like layout and state, is now handled in systems. 
+- Most of the functionality, like layout and state, is now handled in systems.
 
 ### Removed
+
 - Removed `Context` object, use `Resources` instead.