+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added repeat syntax (`hstack![widget;10]`) to `hstack!` and `vstack!` macros.
 - Added `on_click` and `on_hover` gestures.
 - Added `TextField` widget and text input.
+- Added borders.
 
 ### Changed